Adapt worker 'geoserver-publish-imagemosaic'
